requirements.txt
streamlit
yfinance
pandas
requests
python-dotenv
openai
